Extract methods to classes.
